Move scheduling UI components to shared module
